rename app.js to app.ts and move google map stuff to different file
